renames Foreground to ScrollyVideoForeground, adds to exports
